Fix and simplify CI setup (#96)
The problem with the documentation setup introduced in PR 94 was that
the actions were improperly specified. That is fixed in this commit.

In addition, the following changes were made to simplify the setup,
bringing it more in line with Example.jl (I actually copied the file
from there and made individual changes) and most other Julia packages:
- The cronjob for running tests has been removed; it doesn't provide
  meaningful benefit atop running on PRs, tags, and pushes to `main`.
- The concurrency specification has been removed. I believe what's
  specified is actually the default behavior.
- `include` mentions a version check and branch protection but neither
  are present (I assume this was copied from elsewhere) so that step has
  been removed.
- The separate `JuliaNightly` CI has been incorporated back into the
  regular CI setup as an `include` step. This avoids duplication.
- <https://github.com/JuliaRegistries/TagBot> provides a template in its
  README which includes a few more items not present in the one here.
  Those have been added.

Hopefully these changes will make things more consistent and easier to
maintain going forward.